+++ x/sequencer/client/cli/tx.go
@@ -1,18 +1,27 @@
 package cli
 
 import (
+	"encoding/json"
 	"fmt"
+	"strconv"
 	"time"
 
 	"github.com/spf13/cobra"
 
 	"github.com/cosmos/cosmos-sdk/client"
-	// "github.com/cosmos/cosmos-sdk/client/flags"
+
 	"github.com/dymensionxyz/dymension/v3/x/sequencer/types"
+
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/client/tx"
+
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 var (
 	DefaultRelativePacketTimeoutTimestamp = uint64((time.Duration(10) * time.Minute).Nanoseconds())
+	_                                     = strconv.Itoa(0)
 )
 
 // GetTxCmd returns the transaction commands for this module
@@ -26,7 +35,83 @@ func GetTxCmd() *cobra.Command {
 	}
 
 	cmd.AddCommand(CmdCreateSequencer())
+	cmd.AddCommand(CmdUnbond())
 	// this line is used by starport scaffolding # 1
+
+	return cmd
+}
+
+func CmdCreateSequencer() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "create-sequencer [pubkey] [rollapp-id] [description] [bond]",
+		Short: "Create a new sequencer for a rollapp",
+		Args:  cobra.ExactArgs(4),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argPubkey := args[0]
+			argRollappId := args[1]
+			bond := args[3]
+			argDescription := new(types.Description)
+			err = json.Unmarshal([]byte(args[2]), argDescription)
+			if err != nil {
+				return err
+			}
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			var pk cryptotypes.PubKey
+			if err := clientCtx.Codec.UnmarshalInterfaceJSON([]byte(argPubkey), &pk); err != nil {
+				return err
+			}
+
+			bondCoin, err := sdk.ParseCoinNormalized(bond)
+			if err != nil {
+				return err
+			}
+
+			msg, err := types.NewMsgCreateSequencer(
+				clientCtx.GetFromAddress().String(),
+				pk,
+				argRollappId,
+				argDescription,
+				bondCoin,
+			)
+			if err != nil {
+				return err
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdUnbond() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "unbond",
+		Short: "Create a new sequencer for a rollapp",
+		Args:  cobra.ExactArgs(0),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgUnbond(
+				clientCtx.GetFromAddress().String(),
+			)
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
 
 	return cmd
 }
